requirements.txt
streamlit
pandas
matplotlib
requests
praw
feedparser
textblob
python-dotenv
schedule
scikit-learn
supabase
python-telegram-bot>=20.0
plotly>=5.0.0
python-dateutil
pycoingecko
joblib
telegram